+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - legal-vote: add option to set protocol timezone ([#338](https://git.opentalk.dev/opentalk/k3k-controller/-/issues/338))
 - controller: add debriefing and kicking multiple users at once ([#350](https://git.opentalk.dev/opentalk/k3k-controller/-/issues/350))
 - controller: always allow one moderator to join a room regardless of participant limit ([#352](https://git.opentalk.dev/opentalk/k3k-controller/-/issues/352))
+- chat: add private chat history ([#327](https://git.opentalk.dev/opentalk/k3k-controller/-/issues/327))
 
 ### Changed